fix(rust): Handle rebalancing ourselves, remove usage of pre/post-rebalance callbacks
As explained in the code comments, this leads to calling rdkafka_assign
twice: once with the invalid offsets (causing errors), and once with the
fixed-up offsets.

Let's replace the entire flow with something of our own. It turns out
that the use of unsafe in the original implementation of
ConsumerContext::rebalance is also entirely unnecessary.