feat(cardinality): Add clickhouse host and port parameters
In our prod systems, there seems to be an issue when cron jobs try to
access other systems using envoy. In order to bypass that, we can use
the dns name of the cluster to reach the clickhouse host. That is
what we do today in the optimize and cleanup jobs as well. So added the
requirement to pass in clickhouse host information as a CLI argument.