feat(metrics): Create new DataSource for running multiple metrics queries
Effectively, the idea of how to run metrics queries with formulas would be to
run a separate subquery for each timeseries requested, and then combine the
results back together from each of these queries.

The `CompositeQuery` class already has a lot of utlities in place for handling
subqueries, so this PR attempts to leverage those existing utilities. This
means modifying the composite pipeline to create a query plan for each of the
subqueries in a CompositeQuery, run each of them, and then use the result
function to rebuild the results.