Avoid double-parsing Payloads
This will now use the intermediate `serde_json::Value` and convert that into the final type that is used.
We can avoid double-parsing the payload that way, however conversion via `serde_json::from_value` seems to be equally as slow as just re-parsing from scratch, so not sure this buys us that much.

Each processor will now take an already parsed value instead of having to parse the value internally, though there is now a bit of macro magic to create `ProcessingFunction` instances that include the parsing.