feat(profiling): Add ondemand profiler code (#5235)
* feat(profiling): Add ondemand profiler code

setting 'ondemand_profiler_hostnames' to a pod name will now capture a
profile and send it to sentry once the option is un-set again.

we should turn on profiling for all web workers by default (even at low
sample rate) to get a feeling for the overhead. overhead is not linear
with sample rate. there is a baseline cost of having a non-zero value,
and then there is a cost per captured profile.

once there is any sample rate > 0 deployed in Snuba for some pods,
ondemand profiles can be performed for those pods. those are forced into
100% sampling regardless of sample rate (again, for as long as the
sample rate is > 0)

* force profile flush after 30s

* fix typing

* fix typing

---------

Co-authored-by: Meredith Heller <meredith@sentry.io>