fix(ci): Use ghcr for self-hosted image
trying to mimick: https://github.com/getsentry/sentry/pull/61434/files

we already publish nightlies and whatnot to ghcr to serve the needs of
'sentry devservices'. it should be possible to use the same exact
multiarch image for self-hosted.

the rest in cloudbuild.yml seems to run unittests that we already run in
CI