fix(devenv): Make post-merge hook exactly like sentry's
I was a bit annoyed that post-merge in snuba printed multiple commands
on separate lines, so I copied what sentry did. Now instead of this:

    ...
    make install-python-dependencies
    ...
    make install-rs-dev

We print this:

    make install-py-dev install-rs-dev

This is useful to me personally because it's easier to copypaste one
line of terminal scrollback vs many.

install-py-dev is an alias for install-python-dependencies that is also
named consistently with sentry.

Also added apply-migrations, which again, is the same name we use in
sentry.

A new feature is to automatically run those commands on merge, something
that sentry apparently has. I don't know if anybody uses that, i just
copied it over.