fix(capman): escape delimiter characters in allocation policy (#5377)
fixes: https://getsentry.atlassian.net/browse/SNS-2589

Basically we have some custom serialization for the complex configurations in the allocation policy. However it breaks when the actual configuration has a `.`, `:`, or `,` in it and it becomes unserializable. 

The better thing to do would be to change serialization to JSON however that would require a migration of the old keys to the new format which is not an urgent matter.