feat(schemas): Add static validation of rust structs against schema
Add a test that ensures the struct used by the querylog message
processor matches the schema definition.

This is much more flexible than generating structs with typify because
the resulting checks are much more lax, and also because differences can
be ignored on a case by case basis.

On the other hand the boilerplate of having to write Deserialize structs
manually will remain.

The test currently fails because the querylog schema has to be updated.