fix: Force a specific rust toolchain
Snuba contributors who just want a working consumer and are less
interested in their Rust "devenv" are confronted with random compile
errors because their toolchain is outdated by a few months.

This commit forces the same rust toolchain version for everybody. If
somebody runs `cargo build`, rustup will automatically download and
install the right rust version if it does not already exist.

We don't actually want to pin the exact version -- defining an
oldest-supported version is what we actually need, we don't care if the
default toolchain is newer. However, there's no way to declare that,
defining MSRV in Cargo.toml doesn't provide the same experience of
automatically downloading the right version.