fix(capman): register statistical detectors cross org query referrer, make overrides of unregistered referrers not possible (#5422)
I turned on the cross org query policy and even though I had an override for the statistical detectors in it, the policy did not pick it up because it was not a registered referrer. This took too long to understand. This PR:

https://sentry.sentry.io/issues/4889272228/?project=1

* registered the referrer that was being overthrottled
* Makes it so unregistered referrers cannot be overridden at runtime (and tells the user what to do to make that happen)
* Updates the policy decision explanation to tell the user why what is happening is happening

I considered making it possible to override unregistered referrers but I'm being especially protective of this class of queries. They are new to sentry and I don't want people making a whole bunch of them without having us know what they are and what they are for