fix(optimize): Add parallel optimize job cutoff time
When we introduced the parallel optimize job cutoff time, we made its value
same as the overall optimize job cutoff time. As of April 2023, the optimize
job would finish by 3am since the size of individual partitions was no more
than 110GB. Today, the size of each partition is close to 180GB. With the
bigger partitions, it takes longer for the optimize job to finish. Cap the
amount of time the optimize job can run in parallel so that parallel optimizations
are not running during peak hours.

Also, removed the jitter because I don't believe they are adding any value.
Removing it would allow the parallel phase to run a bit longer during the
starting phase of the job.