feat(rust): Implement replacements
- Message processors can now either return an InsertBatch or ReplacementData
- There is a new `ProduceReplacements` strategy which is chained immediately
after the transform step. This strategy forwards inserts to the next step
and produces the replacements. Subsequent steps in the pipeline are unchanged.
- There are currently no message processors that actually produce replacements
since this is only relevant for the "errors" dataset
- This implementation affects the throughput on all message processors,
(based on cargo bench), it may be better to find an implementation that skips
the replacements step entirely for all message processors except errors