fix: Upgrading to Clickhouse 22.8 (#5445)
fix: Upgrading to Clickhouse 22.8

There is a change in Clickhouse >=22.8 where a query can't merge an aggregate state with parameters different from what is stored in the state. For example, if the state is `histogram(250)(value)`, then a query can't do `histogramMerge(100)(value)`. Add a processor that finds instances of this and corrects it to avoid errors in current Sentry use.

This won't have an impact on the product since currently the product is not using the histogram field for metrics.