ref(ddm): Dual write to Sentry metrics as well as Datadog (#5474)
This PR wraps the Datadog and Sentry metrics into a single backend, so we can
start dogfooding Sentry metrics. Currently the Sentry portion is feature
flagged and has a sampling rate.

Sentry also provides a way to specify the unit and record distributions, so add
that to the abstract so we can start using it later.