feat(mql): Parse the entity from the MRI (#5501)
* feat(mql): Parse the entity from the MRI

This has two changes: firstly, resolve public names in queries during the
parsing process, to ensure that we always are dealing with MRIs further on in
the pipeline.

Second, use that restriction to fetch the Entity based on the MRI being passed
in. Our entities align with the metric types, so we can map the metric type of
the MRI to the entity on our end. This will need to be expanded when we start
supporting multi type queries.

This is the first step in not needing an `entity` field in the SDK at all,
which will simplify the SDK and the Sentry layer, and enable features like
using MQL strings directly.