ref(rust): Make errors better for errors processor
The payloads in the errors processor are so large that they're
impossible to debug, so add another crate that keeps track of where a
deserialization error occurs.

We might want to do this for other processors, but the performance
implications of this are unclear, so I'd want to avoid it for payloads
which are already small and easily debuggable.