feat: Remove query splitters from the API
After testing how much the splitters were being used, and the improvement they
were providing, splitters were only being used on a very small percentage
of queries (up to 5% on the highest dataset). After pausing the splitters,
there was almost no impact on the performance of queries. Only a single
referrer had a noticeable performance degradation, and that was not enough to
cause any actual impact on the queries (rate limiting, timeouts etc.)