ref(mql): Return single quantiles as a value instead of an array (#5592)
When fetching quantiles, the results would always be returned as an array, even
if the call only was fetching a single quantile (which is most of the cases).
This was both confusing to the frontend, and was causing errors with Formulas,
since a quantile would return an array, which generally can't be used in math.
(e.g. [1] / 1 is not valid).

When a single quantiles is being fetched, wrap it in arrayElement(..., 1) so
that it unwraps the array and returns a simple value.