feat: Add use_case_id index to generic metrics
The use_case_id is being sent with every query, and in the cases where a
metric_id is not specified, can be used to very effectively reduce the number
of rows scanned.

The granularity is set a little higher than for the tags hash maps, since the
cardinality of use case ID is quite high and the data should be fairly clumpy
(metric_id should all be stored together and each metric_id has only one use
case ID).