feat(metrics): Add a meta table to counters
This creates a table and a materialized view to populate it for storing meta
information about metrics. This table is meant to satisfy queries that are
trying to find metric_ids, tag keys and tag values, but are not interested in
the values associated with the metrics.

In theory this will eventually be done for all the metric types, but for now
this is being used just for counters to test how well this solution actually
solves the problems.