ref(ddm): Add debug queries to cardinality analyzer
For the DDM launch, we want a way to analyze the size of the incoming data. Add
two queries to the predefined queries that can give that information (and
provide a useful starting point for more in depth analysis).

Also change the font of table results to be monospace, so it's easier to
compare numbers directly.