feat(devexp): Separate query processing from query execution stage (#5668)
With this change, we can run two different query pipelines:

1. The old one (default in everything but tests)
2. The new one (using the new QueryPipelineStage interface)

The new query pipeline *does not* depend on the query runner to process the query and the query execution does not need anything except for the physical query (sort of, the `SnubaQueryMetadata` side effects are still there). 

The is mostly wrapping the old stuff with a new wrapper however there is a fundamental split in how this is executed now. 

Main new functionality added:
* Cluster to read from is now selected from the physical query (legacy pipeline uses the logical query)
    * this is necessary to query storages on their own


# Rollout plan

This is a delicate change and therefore requires precision. When this PR is merged, *nothing changes* except that we're passing the `storage_key` with the physical query (a pretty harmless change). 

Once deployed:

(In runtime config)
```
SET run_new_query_pipeline 0.01 # 1 % of queries go through new pipeline
```
... monitor sentry, datadog

Continue increasing the `run_new_query_pipeline` until we hit `1`

Once that woirks, we can remove all of the old pipeline in `web/query`