feat(metrics): Create a table for finding meta tag values
In the original meta table, having tag values in a groupUniqArray had very poor
query performance, driven by the slowness in aggregating those data sketches.

Create a new table that can just be used for finding tag values, that stores
the tag values as rows in the table.

This table also deliberately does not have org_id or use_case_id in it. In
theory any query for tag values has a project ID, metric ID and tag key in it.
This means the org ID and use case ID don't provide anything useful.