feat(meta): Add record_meta columns to metrics tables
This column is used to determine which data should be added to the meta tables,
particularly so that use cases that don't need meta data don't get added
unnecessarily.

This column was added to the counters table as part of the initial experiment,
so it's just the other tables that need it added.