feat(capman): Long term rejection allocation policy (#5718)
Currently we throttle on bytes_scanned but never reject. This new policy (which is only needed for errors atm) keeps track of a ten minute sliding window for all (project|organization, referrer) combinations. This makes it a much more memory heavy process but we have plenty of redis headroom atm.