feat(meta): Adjust partitioning/settings of counters meta tables (#5784)
There was an issue with the existing tables that caused them to have a very
high number of inactive parts. This change makes three adjustments to help with
that.

- Remove `retention_days` from the `PARTITION BY`. This will reduce the number
  of partitions that the table has, and thus the number of inactive parts
- Change the materialized view to use `toMonday` instead of `toStartOfWeek` so
  that the partition lines up with the data in the table. This should help
prevent writing new data into older partitions (which was happening  because
`toStartOfWeek` and `toMonday` don't line up.
- Since partitions are no longer broken up by retention days,
  add the `ttl_only_drop_parts` setting so that we drop rows within the
partitions based on retention.

Since it's not possible to modify the `PARTITION BY` of a table in place, drop
the existing tables and recreate them with the accurate `PARTITION BY`.