fix(meta): Add record_meta field to metrics processing (#5746)
The `record_meta` field is used to decide whether a use case should have meta
information tracked for it. Meta information is used to search metric IDs, tag
keys and tag values, and not all use cases need that information in a
performant way. This was previously just used in counters as testing,  but now
move it out to be used by all the metric types.