fix(22.8) Add processor to handle 'allocator' bug in Clickhouse (#5802)
* fix(22.8) Add processor to handle 'allocator' bug in Clickhouse

There is some bug in Clickhouse 22.8 that can cause errors like "Too large size
(13837309855095848992) passed to allocator. It indicates an error."

The issue was narrowed down to conditions on low cardinality string columns,
and specific types of data in those columns. This processor hacks those
conditions to be in a form that allows the query to run.

* remove print

* add flag

* fix has array bug

* remove print

* restrict to only LowCardinality columns

* add version

* stop running on 21.8

* test values not working

* add test

* remove optimizer from discover

* fix faulty test selection

---------

Co-authored-by: Volo Kluev <volo.kluev@sentry.io>