fix(22.8) Cast LowCardinality string columns to String
In 22.8, if a LowCardinality(String) is queried, and it actually is not low
cardinality, it can cause an error in Clickhouse. To be safe, cast all
LowCardinality columns to regular String to avoid causing this issue.

So far this is only impacting Discover + transactions, so limit the change
there. This might decrease query performance for other queries, since
all LowCardinality strings will be treated as String.