fix(meta) Rename counter tables for consistency (#5797)
These were created with new names to avoid conflicting with the old tables when
the partition keys were changed. However for consistency's sake it would be
nice to have all the tables in the different clusters have the same naming
schema.

Drop the old tables and recreate them with the correct names.