feat(meta) Add configuration for meta tables in generic metrics (#5824)
This adds all the YAML files for the storages and entities required to query
the meta tables. Note that this is 4 copies of the same files, just replacing
the metric type in each one (counters/sets/gauges/distributions).

A required column for the meta table is the use case ID, but that is a string
column. This also slightly modifies the validator to allow string columns as
well as integer columns. The intent was to keep the strictness of the original
check (which specifically looks for ints), so this added a new similar check.