feat(dsl) Add a visitor that maps AST objects into DSL strings (#5844)
Most of the AST objects that exist in the tests have a DSL syntactic sugar
function to go with them. This code will take AST objects as input, and output
a string that is valid Python code for the corresponding DSL function.

This is a rough MVP of this, and could use some polishing/cleanup.