ref: Deprecated TOPIC_PARTITION_COUNTS setting
Snuba should fetch the actual value from the broker on startup rather than
rely on a hardcoded TOPIC_PARTITION_COUNTS settings. This will ensure that
it always has the correct values and remove the manual process and associated
drift in ops that is currently involved in keeping this up to date.