fix: Handle tag columns in low cardinality processor (#5875)
Discover will sometimes send filters on columns as tags instead of actual
columns, e.g. tags[environment]. This was causing the allocator bug to be hit
again on some queries. This change handles checking for this types of queries
and correctly casting them as well to avoid the allocator bugs.