ref: Deprecate TOPIC_PARTITION_COUNTS setting (#5868)
Snuba should fetch the actual value from the broker on startup rather than rely on a hardcoded TOPIC_PARTITION_COUNTS settings. This will ensure that it always has the correct values and remove the manual process and associated drift in ops that is currently involved in keeping this up to date.

Depends on https://github.com/getsentry/sentry/pull/70478

This is a prerequisite to https://github.com/getsentry/ops/pull/10392 and https://github.com/getsentry/ops/pull/10451 which will enable clean up of topicctl configuration in ops.