fix(meta) Change the view to use LEFT ARRAY JOIN
In the current implementation if a metric has no tags associated with it, it
will get excluded from the meta table since the array join only selects rows
that have values.

This change will cause rows to be created with an empty tag key (0) so that the
metrics can still be searched for in the meta tables.