ref(subscriptions): Add query-based exception handling in scheduler  (#5976)
Currently, encoding a subscription task can raise an InvalidQueryException if there is a malformed subscription query being built that fails query processing (e.g. one that is missing an added column condition). The current behavior leads to the consumer pods to crashloop, meaning all subscriptions are delayed if a single query is malformed.

Instead, we can handle the exception and simply skip the offset for the malformed query. We're trying this in this PR. Might add a test to this too. I also wonder if this is doable given the way the scheduler needs to track ticks and the offsets against them.


---------

Co-authored-by: Evan Hicks <evan.hicks@sentry.io>