ref(spans): Create script for generating fake traces (#6032)
Create a script that generates fake traces for a set of projects.

All the traces for a single organization (1123) that spans 3 projects. Conceptually it is meant to
mimic a frontend, backend and service project.

The script can write to a file, output to stdout or produce to a kafka topic. It's all random, so
there are no guarantees of how many transactions/spans will actually appear in a given trace. There
is an option to run the script until a given number of spans are generated. Note this isn't exact,
the script will run until at least that many spans are generated.