feat(profiling): Add profiler id column to transactions (#6099) (#6103)
Re-applying #6099.

For continuous profiling, the profile context will contain a new value
`profiler_id`, not to be confused with `profile_id`. A `profiler_id` is set on
the transaction's profile context when the continuous profiler is enabled. This
is a separate value from the existing `profile_id` because we need to be able to
distinguish if the transaction is associated with a transaction based profile or
a continuous profile and load the profile from the correct place.