ref(lwdeletes): create parse_and_run_query function (#6120)
### Context

Lightweight deletes needs to run a query to check how many rows will be affected before actually running the deletes. This requires that the code to run a query is reusable. There is also a disconnect in the codebase between the usefulness of datasets.

The original intent in snuba was that datasets where central to how things were queried. That was never really realized. In SnQL queries, entities or storages are specified directly. There is no functional need for datasets. However, many metrics are tagged with the dataset name and many dashboards use it as filters. This means we can't remove datasets entirely just yet.

### What this PR does

Creates a `parse_and_run_query` function which is the new top level function for running queries from the input JSON payload. It does not concern itself with any HTTP error handling, it just runs the query pipeline. Can be used in a modular way (necessary for lightweight deletes)