feat: new page in snuba admin to do deletes (#6168)
this is a redo of #6157

This PR adds a page to snuba admin that we can use to test deletes in
prod.

## major changes
1. create a front-end react interface for deletes
* `static/delete_tool/index.tsx` is the new react UI, it calls
`api_client.tsx:runLightweightDelete`
* `api_client.py:runLightweightDelete` sends the input via http request
to `/delete` endpoint. (note: i think it would have been equally valid
to just send the http request from index.tsx rather than having this
separate function, but I did it like this bc everyone else in the
codebase did)
* `data.tsx` adds the new page to the Nav bar
2. create a new `/delete` endpoint in snuba admin
`admin/views.py:delete`
As said in bullet 1, the react UI sends the input to `/delete`. This is
a new endpoint we created, it mostly just calls `delete_from_storage`
(what Meredith implemented that actually does the delete)
3. set permissions for `/delete` endpoint and UI `tool_policies.py`
It just goes into AllTools by default which is the most restrictive
role, only snuba team and people with access to everything can use it. I
think this should be good.
4. small `delete_query.py` changes
* I slightly changed the input/output types of
`delete_query.py:delete_from_storage` in order to make the interface
more strict/clear.
* added more input verification for the `columns` input.

## considerations
This tool is pretty much just a layer on top of `delete_from_storage`
meaning it can do anything that `delete_from_storage` can do. Hence it
will be able to delete in any environment, and any dataset that has
deletes enabled. Is this fine or should we add additional strictness to
this interface, so that it can only be used for search_issues in s4s no
matter what.

<img width="1212" alt="Screenshot 2024-07-30 at 1 10 29 PM"
src="https://github.com/user-attachments/assets/d3ad9005-69b1-4995-8934-f85494fdb3b5">