feat(eap): Add attribute key/value lookup table (#6152)
Create a table that stores just the attribute keys and values that have
been seen by an org. This
can be used for autocomplete or any other validation feature.

The table doesn't have any indexes on it at the moment. It also provides
a count, to do simple
ranking if that's useful in the future.