ref(snuba-deletes): introduce delete allocation policies (#6180)
**context**
Right now we have a simple ratelimiter that limits deletes to `1`
concurrent query at a time for the whole endpoint
https://github.com/getsentry/snuba/blob/ea2f8e716ea519ae173e7a6833da5d2bf16859f2/snuba/web/views.py#L305

This is good enough for testing out ourselves, but beyond that we wanted
to add a more robust ratelimiter that allowed for more granular levers
that we have for other endpoints. Product should also be ratelimiting on
their end though, this is not supposed to be in lieu of that.

**delete allocation policies**
We can reuse the allocation policy framework for the delete query
pipeline. I've made these policies separate to the other allocation
policies definitions because we should be adjusting the limits
separately; deletes should not have as high thresholds as select queries
because the load those queries cause will be higher, even though they
are 'lightweight' (lightweight compared to heavy mutations).

**`DeleteConcurrentRateLimitAllocationPolicy`**
I made a separate class because I want the redis key to be different
than the `ConcurrentRateLimitAllocationPolicy` so that we don't clobber
ratelimits

This PR just adds the set up for the allocation policies, and does
nothing with them (aside from adding they to the `Query` object. The
rest of the logic will be in
https://github.com/getsentry/snuba/pull/6181