Increase the default throttle/warning threshold for allocation policies (#6189)
Allocation policies are our mechanism for doing traffic management for
Snuba queries. Currently, we see a lot of warnings (Warning: Query from
referrer ... is throttled) on Sentry Issues. This is because ~17% of
queries to Snuba are throttled by capacity management/allocation
policies. This is an overkill since we actually have a lot of Clickhouse
capacity.

This PR increases the default throttle/warning threshold to be higher
than half of the rejection threshold. I will also increase the
thresholds in Snuba Admin.

Co-authored-by: Rachel Chen <rachelchen@PL6VFX9HP4.local>