ref(snuba-deletes): add ColumnValidator (#6198)
We have some [schema
validation](https://github.com/getsentry/snuba/blob/d4b164c06710e15bed33d868102305adf8fe495c/snuba/query/schema.py#L38)
on delete api, but because the columns depend on the storage, we can't
do that level of type validation when parsing the payload. The only
required column we have is the `project_id` (which is debatable on
whether we should have that there but anyway)

If the column `occurrence_id` is an uuid and someone sends:

```json
{ "columns": {"occurrence_id": ["hello"]}
```
we'll get a clickhouse error when trying to run the delete.

**ColumnValidator**
This validator will check to make sure that the type defined in the
schema matches the values passed through in the api. I've only
implemented a handful of types and I don't expect that we will need many
more, at least for now.