bug(mql): Fix WITH TOTALS for MQL queries (#6218)
### Overview
This PR is responsible for fixing a bug with the `totals` flag in MQL.
When the `totals` flag is set to true, snuba will insert a `WITH TOTALS`
modifier next to the `GROUP BY` in the query. Read more about WITH
TOTALS
[here](https://clickhouse.com/docs/en/sql-reference/statements/select/group-by#with-totals-modifier).
However, if the query does not contain a `GROUP BY` then the modifier
will not be added to the query. As a result, the flag is ignored and CH
does not return a totals row. When this happens, the query AST still has
`totals` set to true, and during [result
transformation](https://github.com/getsentry/snuba/blob/fc11af17f7558e56a4139d061f47bef0369de0b7/snuba/clickhouse/native.py#L472),
snuba pops the last element of the data array and assumes it is the
totals row (even though it isn't).

To fix this, in MQL parsing, we should only set and respect the totals
flag only if there exists a group by. If not `WITH TOTALS` cannot be
applied anyways, and therefore we should not set the flag.