ref(snuba-deletes): use allocation policies (#6181)
**context**
We have the configurations for the delete allocation policies but they
aren't actually being used in the delete pipeline yet.
Some more info in the original PR that added the configuration for the
delete allocation policies https://github.com/getsentry/snuba/pull/6180.

This reuses some of the functionality from the regular select query
pipeline. The delete concurrent policy we defined will not be enforced
yet, so we will still be relying on the simple rate limiter
https://github.com/getsentry/snuba/blob/ea2f8e716ea519ae173e7a6833da5d2bf16859f2/snuba/web/views.py#L305
until we are ready to enforce the allocation policy. We can also add
another delete policy once we verify the delete allocation policy
pipeline is working as expected.

**next steps**
- [ ] add delete allocation policies to the snuba admin UI so that we
can change values https://github.com/getsentry/snuba/pull/6248
- [ ] test out the delete allocation policy to verify it
- [ ] add a delete referrer guard policy