Emits DD metric for queries statuses (#6261)
We currently emit these metrics on a per-policy level, which means some
queries can be counted multiple times for multiple policies. This PR
adds metrics so we only count each query once. This will allow us to get
a more accurate sense of the % of queries that are
successful/throttled/rejected

Co-authored-by: Rachel Chen <rachelchen@PL6VFX9HP4.local>