cleanup(capman): remove legacy table rate limits (#6274)
Before the capacity management framework we would apply rate limits as
query processors in the query pipeline. That is no longer necessary. The
TableRateLimit was the last holdout of these legacy rate limits but it
is no longer necessary for these reasons:

1. The existing policies cover all the necessary rate limiting and
throttling under normal operations
2. When we lose a clickhouse replica or the cluster capacity diminished
drastically, it is the clickhouse nodes themselves which reject the
query, not the table rate limit.

Functionally, this table rate limit is wasted code and CPU cycles