feat(job-runner): scaffolding for job manifest testing (#6282)
In this PR, I'm creating a sample job run manifest along with a reader
and an API endpoint to return what's in the static file. This will help
me test whether the plumbing to put different static files in this
location is working.

Next steps: deploy to prod, validate the correct result. Start working
on providing a different payload in a different environment based on the
ops repo.