Cleanup some settings from settings_test.py (#6313)
During the investigation of
https://github.com/getsentry/snuba/pull/6279, I found that:
- `OPTIMIZE_JOB_CUTOFF_TIME` is overridden to `24` and it was causing
test failures.
- `OPTIMIZE_PARALLEL_MAX_JITTER_MINUTES` does not seem to be used
anywhere in the code.

With the change, less no. of local tests fail. The failure number is
non-zero because https://github.com/getsentry/snuba/pull/6279 was
reverted yesterday . I would like to verify if this small change passes
through the CI and stays stable in prod.