feat(eap): Include server-side sample rates in sampling weight (#6333)
Adds support for the new `server_sample_rate` measurement on spans and
includes it in computing the final sampling factor and weight. This
allows us to extrapolate span metrics correctly in organizations that
have dynamic sampling enabled.

This also rounds the sampling factor to a precision of 10^-9 to avoid
floating point errors in the multiplication. For example, without the
test using sample rates of _10%_ and _20%_ respectively already shows
floating point error.

In a follow-up, we might revise the schema and move these sample rates
into `span.data`.