feat(database-clusters): Implement V1 of the database clusters tab in Snuba Admin (#6345)
Implemented V1 of the database clusters tab in Snuba Admin. The purpose
of this tab is to show the nodes on which clickhouse is running as well
as their versions.

Future work: Instead of using a raw table, there should be groupings and
relationships that allow us to better visualize the cluster setup (e.g.
clusters should be grouped together). We should also add more metadata
so that it's easier to understand the topology of our clusters
(distributed nodes should be differentiated from regular storage nodes).

<img width="1462" alt="image"
src="https://github.com/user-attachments/assets/5b9d2331-1c37-4ac7-8da5-8b188215f02d">