try to fix failed migration events_analytics_platform, 0001_spans (#6356)
fix for syntax error in migration :
Codec DoubleDelta does not accept any arguments.

https://github.com/getsentry/snuba/issues/6355

<!--

  Sentry employees and contractors can delete or ignore the following.

-->

### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.

---------

Co-authored-by: iops <iops@syneforge.com>