feat(api): Add a clickhouse connection sanity check to api healthcheck (#6408)
### Overview

This is a follow up item for inc-894, where a API pod landed on a bad
node with no clickhouse connection. Because it was not configured with a
thorough healthcheck, the pod stayed and served 5XX to all requests.

Adds a connection check to see if any clickhouse query node is operable.
Fails the healthcheck if all query nodes are inoperable, as it likely
indicates connection error with the host, so k8 can rotate the pod.
Renames the previous check_clickhouse function to
`check_all_tables_present` to distinguish the 2.