ref(eap): Implement synchronizer
Induce a hardcoded delay so that the eap-mutations consumer will always
lag behind the eap-spans consumer.

This PR (and most of my work on the mutations consumer) makes a point to not use custom strategies, as that's too much
boilerplate. Rather, we should improve the arroyo API. For this PR, this
arroyo PR is necessary: https://github.com/getsentry/arroyo/pull/388