ref(querylog): Produce request ID consistently
There are two ways to stringify a UUID: `str(uuid)`, and `uuid.hex`

It seems we use both inconsistently, and the consumer can only deal with
the latter.

But also, sometimes we do this: `str(uuid.uuid4)`

That always produces something like `'<function uuid4 at 0x100c39620>'`,
which is total garbage but since everything is string-typed, mypy
doesn't complain.

Refactor everything to use UUID types and only convert to string on
serialization.