Fix snuba admin's query tracing to connect to right storage and query nodes (#6268)
Snuba admin's tracing tool shows query trace output. In addition to
that, it will now show the profile events data. This PR parses the trace
output and builds a `dict` of node name to query id. Then system query
is executed on those nodes to get the profile events for the
corresponding query id.

Before connecting to the nodes, a socket connection is attempted to see
if the hostname resolves. If it does not, then the node is assumed to be
`127.0.0.1`. This is required because while running locally, snuba admin
process will connect to clickhouse container. But the container's
hostname does not resolve, defaulting to `127.0.0.1 `as the hostname.

To make sure hostnames in the CI jobs resolve, the CI docker-compose
file is updated.

---------

Co-authored-by: David Tsukernik <david.tsukernik@sentry.io>