add __eq__ definition to the MigrationModifiers Dataclass
This is a temporary fix to overcome the issue where comparison of
modifiers between the storage config definition and the actual schema
definition fail due to different type of modifier classes.

The modifiers from the storage config are instantiated as
SchemaModifiers, whereas the modifiers parsed from the actual table
schema are instantiated as MigrationModifiers.