ref(bulk-deletes): bulk delete query path and producer (#6440)
**context**
The current pipeline we have for deleting data (via lightweight deletes)
is an API endpoint that executes the query directly. `DELETE` requests
are technically synchronous at the moment due to the ClickHouse version
we are on (the `lightweight_delete_mutations_sync` parameter exist in
newer version).

**problem**
We'd like to keep the number of concurrent `DELETE` requests to a
minimum since they add an increased load to ClickHouse. Though the
number of rows being deleted does have an effect on load, even a delete
that deletes a few rows can have a significant enough effect because
it's still a mutation (under the hood the `DELETE FROM` becomes an
`UPDATE`).

**batched deletes**
In order to scale we are adding a consumer that will batch up DELETE
queries so that we have fewer mutations, even though the number of rows
deleted/mutated will be higher.

The new pipeline is a bit different and has some new components:
- [ ] `DELETE` api endpoint will now take a request and produce it to a
topic (This PR)
* The topics are storage specific, ex. `LW_DELETIONS_SEARCH_ISSUES =
"snuba-lw-deletions-search-issues"`
- [ ] Consumer(s) will be added to consume from the topic that
corresponds to the storage the consumer is for, right now this will just
be search issues

**other notable changes (in this pr)**
* addition of `allowed_columns` in the deletion settings, this will
eventually replace the `ColumnFilterProcessor` once deletes are fully on
the bulk delete endpoint
* `rows_to_delete` is added to the delete query payload so that we can
batch based on number of rows to be delete instead of batching on number
of messages